+++ src/handlers/guildDelete.js
@@ -1,6 +1,11 @@
 exports.handle = function (guild) {
-  this.ddog.increment('event.guildDelete')
-  this.ddog.decrement('total.guildsGained')
-  this.db.deleteGuild(guild.id)
-  this.db.deleteDevSubscriber(guild.id)
-}
+  if (guild.unavailable) {
+    this.ddog.increment('event.guildunavailable');
+    return;
+  }
+  this.stats.guildsLeft++;
+  this.ddog.increment('event.guildDelete');
+  this.ddog.decrement('total.guildsGained');
+  this.db.deleteGuild(guild.id);
+  this.db.deleteDevSubscriber(guild.id);
+};
